Moving to the undeprecated Preview API for Operational Insights
